+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.0 (2019-02-24)
+
+-   Added: `foldersIds` option
+-   Added: `fields` option
+-   Added: `fieldsMapper` option
+-   Removed: `documents` option
+-   Updated: Default permissions to read document from Google Drive folders
+
 # 0.1.0 (2019-02-07)
 
 -   Added: Headings support